refactor(lib): update imports for new structure
