+++ src/app/sitemap.ts
@@ -1,6 +1,6 @@
 import { MetadataRoute } from "next";
 import { client } from "@/src/sanity/lib/client";
-import { allProgramsQuery } from "@/src/lib/queries";
+import { allProgramsQuery } from "@/src/lib/sanity/queries";
 import { Program, CDKey } from "@/src/types";
 
 export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
