+++ src/components/program/cdkeys/ReportPopup.tsx
@@ -3,18 +3,23 @@
 import { useState, useEffect, useCallback } from "react";
 import { createPortal } from "react-dom";
 import { CDKey } from "@/src/types/program";
-import { trackEvent } from "@/src/lib/trackEvent";
+import { trackEvent } from "@/src/lib/analytics/trackEvent";
 import { DuplicateCheckRequest, DuplicateCheckResponse } from "@/src/types";
 import Toast from "@/src/components/ui/Toast";
 import RenewalModal from "./RenewalModal";
 import { FiCheck, FiX, FiAlertTriangle, FiRefreshCw } from "react-icons/fi";
-import { EVENT_TYPE_MAP, getReportButtonConfig, getStatusTextFromEventType, CDKeyStatus } from "@/src/lib/cdKeyUtils";
+import {
+  EVENT_TYPE_MAP,
+  getReportButtonConfig,
+  getStatusTextFromEventType,
+  CDKeyStatus
+} from "@/src/lib/program/cdKeyUtils";
 import {
   NOTIFICATION_DURATION,
   getReportStatusMessage,
   getErrorMessage,
   getInfoMessage
-} from "@/src/lib/notificationUtils";
+} from "@/src/lib/notifications/notificationUtils";
 import { formatDate } from "@/src/lib/dateUtils";
 
 interface ReportPopupProps {
